requirements.txt
flask~=2.0.1
flask-pymongo
pymongo~=3.11.4
Werkzeug~=2.0.1
flask-restful
PyJWT==2.1.0
cryptography==3.4.7
passlib~=1.7.4